# frontend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: git-agent HTTP API
  version: 0.1.0
  description: >
    The front↔back contract for git-agent. This is the authoritative wire spec:
    the frontend (src/api/contract.ts + http.ts) and the future FastAPI backend
    both implement it. The backend is a thin HTTP layer over the existing
    `core/runtime/runtime.py::Runtime` facade plus the sub-agent system
    (`core/agents/subagents/`). Schema field names are camelCase on the wire;
    the mapping to Python symbols is noted per schema. SSE event payloads are in
    components.schemas.RunEvent.

    Security: `llm_api_key` is stored plaintext in the `runs` table and returned
    by `RunStore.get`. The HTTP layer MUST redact it — only `keyMasked` ever
    crosses the wire. Idempotency: POST /runs is keyed on (repo, commit, model);
    the response carries a `disposition`.
servers:
  - url: /api
paths:
  /runs:
    get:
      summary: List runs
      operationId: listRuns
      responses:
        "200":
          description: All runs
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items: { $ref: "#/components/schemas/Run" }
    post:
      summary: Submit a run (idempotent per repo+commit+model → Runtime.submit)
      operationId: submitRun
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitRunRequest" }
      responses:
        "200":
          description: Created/resumed/attached/already-succeeded run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmitRunResponse" }
        "409":
          description: Active run held by a live lease (Runtime ConflictError)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /runs/{id}:
    get:
      summary: Get one run (RunStore.get — redact the key)
      operationId: getRun
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: The run
          content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /runs/{id}/cancel:
    post:
      summary: Request cancel (Runtime.cancel → cancel_requested_at mailbox)
      operationId: cancelRun
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: The run after the cancel request
          content: { application/json: { schema: { $ref: "#/components/schemas/Run" } } }
  /runs/{id}/resume:
    post:
      summary: Resume a failed/interrupted run (resubmit → disposition "resumed")
      operationId: resumeRun
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: The resumed run + disposition
          content: { application/json: { schema: { $ref: "#/components/schemas/SubmitRunResponse" } } }
  /runs/{id}/report:
    get:
      summary: Get the run's report (runs.report JSONB — nodes.py::report shape)
      operationId: getReport
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: The report
          content: { application/json: { schema: { $ref: "#/components/schemas/Report" } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /runs/{id}/graph:
    get:
      summary: The run's node graph (pipeline nodes and/or lead+sub-agent star)
      operationId: getGraph
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: Nodes + edges
          content: { application/json: { schema: { $ref: "#/components/schemas/RunGraph" } } }
  /runs/{id}/nodes/{nodeId}:
    get:
      summary: Node spec (procedural node, lead, or sub-agent delegation)
      operationId: getNodeSpec
      parameters:
        - { $ref: "#/components/parameters/RunId" }
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The node spec
          content: { application/json: { schema: { $ref: "#/components/schemas/NodeSpec" } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /runs/{id}/events:
    get:
      summary: >
        SSE event stream (MemoryStreamBridge replay + live). Each SSE `data:`
        line is one RunEvent. Resume with `cursor`; the backend replays events
        with cursor > the given value, or emits a `gap` event if the cursor fell
        out of the in-memory buffer (client then refetches via Runtime.events).
      operationId: streamRunEvents
      parameters:
        - { $ref: "#/components/parameters/RunId" }
        - name: cursor
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: text/event-stream of RunEvent objects
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/RunEvent" }
  /runs/{id}/chat:
    get:
      summary: Post-run chat history (agent runs) — persisted turns
      operationId: getChatHistory
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: Chat turns
          content:
            application/json:
              schema:
                type: object
                required: [turns]
                properties:
                  turns:
                    type: array
                    items: { $ref: "#/components/schemas/ChatTurn" }
    post:
      summary: >
        Send a message to the lead agent of a finished agent run. Continues the
        run's checkpoint thread (does NOT create a new run or change its status);
        streams the reply as SSE RunEvent frames, then a final {"type":"chat_done"}.
      operationId: sendChat
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties: { message: { type: string } }
      responses:
        "200":
          description: text/event-stream of RunEvent frames
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/RunEvent" }
        "422":
          description: Not an agent run (chat unavailable)
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /connections:
    get:
      summary: List saved LLM connections (NET-NEW table; keys redacted)
      operationId: listConnections
      responses:
        "200":
          description: Connections
          content:
            application/json:
              schema:
                type: object
                required: [connections]
                properties:
                  connections:
                    type: array
                    items: { $ref: "#/components/schemas/Connection" }
    post:
      summary: Create a connection (apiKey write-only)
      operationId: createConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, apiBase, apiKey, model]
              properties:
                name: { type: string }
                apiBase: { type: string }
                apiKey: { type: string, writeOnly: true }
                model: { type: string }
      responses:
        "200":
          description: Created connection
          content: { application/json: { schema: { $ref: "#/components/schemas/Connection" } } }
  /connections/{id}:
    delete:
      summary: Delete a connection
      operationId: deleteConnection
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Deleted }
  /connections/{id}/check:
    post:
      summary: Ping the connection endpoint (GET {apiBase}/models)
      operationId: checkConnection
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Connection with updated lastCheck
          content: { application/json: { schema: { $ref: "#/components/schemas/Connection" } } }
  /sandboxes:
    get:
      summary: List sandboxes (sandboxes table)
      operationId: listSandboxes
      responses:
        "200":
          description: Sandboxes
          content:
            application/json:
              schema:
                type: object
                required: [sandboxes]
                properties:
                  sandboxes:
                    type: array
                    items: { $ref: "#/components/schemas/SandboxSpec" }
    post:
      summary: Create a sandbox (ssh kind NotImplemented in the backend)
      operationId: createSandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind]
              properties:
                name: { type: string }
                kind: { $ref: "#/components/schemas/SandboxKind" }
                image: { type: string }
                workdir: { type: string }
      responses:
        "200":
          description: Created sandbox
          content: { application/json: { schema: { $ref: "#/components/schemas/SandboxSpec" } } }
...
  /sandboxes/instances/{id}/kill:
    post:
      summary: Kill a sandbox instance (destroy + mark dead; idempotent)
      operationId: killSandboxInstance
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The instance, now dead
          content: { application/json: { schema: { $ref: "#/components/schemas/SandboxInstance" } } }
        "404":
          description: Instance not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }

    # ── runs (runs table + core/runtime/schemas.py) ──
    RunStatus:
      type: string
      description: core/runtime/schemas.py::RunStatus. Terminal = succeeded|failed|interrupted (absorbing). Resume failed|interrupted → pending via claim only.
      enum: [pending, running, succeeded, failed, interrupted]
...
    StopReason:
      type: [string, "null"]
      enum: [orphan_recovered, cancelled, shutting_down, null]
...
    Run:
      type: object
      description: A row of the `runs` table + runtime fields.
      required: [id, repositoryId, repoUrl, repo, status, attempt, connection, hasReport, metrics]
      properties:
        id: { type: string }
        repositoryId: { type: string }
        repoUrl: { type: string }
        repo: { type: string, description: "display slug org/name" }
        commitSha: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/RunStatus" }
        error: { type: [string, "null"] }
        stopReason: { $ref: "#/components/schemas/StopReason" }
        cancelRequestedAt: { type: [string, "null"], format: date-time }
        attempt: { type: integer }
        connection: { $ref: "#/components/schemas/RunConnection" }
        sandbox: { type: [string, "null"] }
        memoryPreset: { type: [string, "null"] }
        hasReport: { type: boolean }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, "null"], format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
        metrics: { $ref: "#/components/schemas/RunMetrics" }
    SubmitRunRequest:
      type: object
      required: [repoUrl]
      properties:
        repoUrl: { type: string }
        branch: { type: string }
        mode:
          type: string
          enum: [pipeline, agent]
          description: >
            Run mode: pipeline (scan→parse→report, default) or agent (lead with
            sub-agent delegation). Run identity does not include the mode —
            resubmitting the same (repo, commit, model) joins the existing run.
        instructions:
          type: string
          description: >
            User task for the run (agent mode; "{repo_url}" placeholder is
            substituted). Ignored by the pipeline mode. Applies at run
            creation only — resume continues from the checkpoint.
        connectionId: { type: string, description: "saved connection (NET-NEW)" }
        model: { type: string }
        apiBase: { type: string }
        apiKey: { type: string, writeOnly: true }
        sandbox: { type: string, description: "sandbox name (create_sandbox_by_name)" }
        memoryPreset: { type: string }
        features: { $ref: "#/components/schemas/RunFeatures" }
...
    SubmitRunResponse:
      type: object
      required: [run, disposition]
      properties:
        run: { $ref: "#/components/schemas/Run" }
        disposition: { $ref: "#/components/schemas/SubmitDisposition" }

    # ── report (nodes.py::report) ──
...
    Report:
      type: object
      required: [repoUrl, commit, description, structure, modules, dependencies, skippedFiles]
      properties:
        repoUrl: { type: string }
        commit: { type: string }
        description: { type: string }
        structure: { $ref: "#/components/schemas/ReportStructure" }
        modules: { type: array, items: { $ref: "#/components/schemas/ReportModule" } }
        dependencies: { type: array, items: { type: string } }
        skippedFiles: { type: array, items: { type: string } }
        error: { type: string }

    # ── graph + node spec ──
...
    NodeKind: { type: string, enum: [procedural, agent] }
...
    RunGraph:
      type: object
      required: [runId, nodes, edges]
      properties:
        runId: { type: string }
        nodes: { type: array, items: { $ref: "#/components/schemas/GraphNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/GraphEdge" } }
...
    NodeSpec:
      type: object
      description: >
        Static description of a node. Procedural: systemPrompt = inline prompt or
        null, tools = sandbox commands. Lead: sandbox toolset + `task` tool.
        Sub-agent: registry type (registry.py) with a delegation runtime.
      required: [id, label, kind, description, systemPrompt, tools, model, memoryPreset]
      properties:
        id: { type: string }
        label: { type: string }
        kind: { $ref: "#/components/schemas/NodeKind" }
        description: { type: string }
        systemPrompt: { type: [string, "null"] }
        tools: { type: array, items: { $ref: "#/components/schemas/ToolSpec" } }
        model: { type: [string, "null"] }
        memoryPreset: { type: [string, "null"] }
        subagentType: { type: string }
        maxTurns: { type: integer }
        timeoutSeconds: { type: number }
        delegation:
          oneOf: [{ $ref: "#/components/schemas/TaskDelegation" }, { type: "null" }]

    # ── sub-agents (core/agents/subagents/) ──
...
    Connection:
      type: object
      required: [id, name, apiBase, model, keyMasked, createdAt, lastCheck]
      properties:
        id: { type: string }
        name: { type: string }
        apiBase: { type: string }
        model: { type: string }
        keyMasked: { type: string }
        createdAt: { type: string, format: date-time }
        lastCheck:
          oneOf:
            - type: object
              required: [ok, latencyMs, at]
              properties:
                ok: { type: boolean }
                latencyMs: { type: integer }
                at: { type: string, format: date-time }
            - type: "null"
    SandboxKind: { type: string, enum: [opensandbox, local, ssh] }
    SandboxSpec:
      type: object
      required: [id, name, kind, image, workdir, createdAt, runCount]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { $ref: "#/components/schemas/SandboxKind" }
        image: { type: [string, "null"] }
        workdir: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        runCount: { type: integer }
...
    ChatTurn:
      type: object
      required: [role, text]
      properties:
        role: { type: string, enum: [user, agent] }
        text: { type: string }
...